Muuta AineistoManagerin nimeksi TiedostoManager ja nimeä [vaihe]Aineistot paremmin [Vaihe]TiedostoManagereiksi.
