--- backend/src/aineisto/ProjektiAineistoManager/AbstractHyvaksymisPaatosVaiheAineisto.ts
@@ -1,4 +1,0 @@
-import { VaiheAineisto } from ".";
-import { HyvaksymisPaatosVaihe, HyvaksymisPaatosVaiheJulkaisu } from "../../database/model";
-
-export abstract class AbstractHyvaksymisPaatosVaiheAineisto extends VaiheAineisto<HyvaksymisPaatosVaihe, HyvaksymisPaatosVaiheJulkaisu> {}
