--- backend/src/aineisto/ProjektiAineistoManager/index.ts
@@ -1,13 +1,0 @@
-export * from "./ProjektiAineistoManager";
-export * from "./util";
-export * from "./VaiheAineisto";
-export * from "./AbstractHyvaksymisPaatosVaiheAineisto";
-export * from "./S3Paths";
-export * from "./AineistoPathsPair";
-export * from "./AloitusKuulutusAineisto";
-export * from "./HyvaksymisPaatosVaiheAineisto";
-export * from "./JatkoPaatos1VaiheAineisto";
-export * from "./JatkoPaatos2VaiheAineisto";
-export * from "./NahtavillaoloVaiheAineisto";
-export * from "./VuorovaikutusKierrosAineisto";
-export * from "./VuorovaikutusKierrosJulkaisuAineisto";
