+++ backend/src/aineisto/ProjektiTiedostoManager/AbstractHyvaksymisPaatosVaiheTiedostoManager.ts
@@ -1,0 +1,7 @@
+import { VaiheTiedostoManager } from ".";
+import { HyvaksymisPaatosVaihe, HyvaksymisPaatosVaiheJulkaisu } from "../../database/model";
+
+export abstract class AbstractHyvaksymisPaatosVaiheTiedostoManager extends VaiheTiedostoManager<
+  HyvaksymisPaatosVaihe,
+  HyvaksymisPaatosVaiheJulkaisu
+> {}
