+++ backend/src/handler/tila/vuorovaikutusKierrosTilaManager.ts
@@ -26,7 +26,7 @@ import { isKieliSaame, isKieliTranslatable, KaannettavaKieli } from "hassu-commo
 import SMTPTransport from "nodemailer/lib/smtp-transport";
 import { isOkToMakeNewVuorovaikutusKierros } from "../../util/validation";
 import { yhteystiedotBackToStandardiYhteystiedot } from "../../util/adaptStandardiYhteystiedot";
-import { ProjektiAineistoManager, VaiheAineisto } from "../../aineisto/ProjektiAineistoManager";
+import { ProjektiTiedostoManager, VaiheTiedostoManager } from "../../aineisto/ProjektiTiedostoManager";
 import { examineEmailSentResults, saveEmailAsFile } from "../../email/emailUtil";
 
 class VuorovaikutusKierrosTilaManager extends TilaManager<VuorovaikutusKierros, VuorovaikutusKierrosJulkaisu> {
@@ -322,8 +322,8 @@ class VuorovaikutusKierrosTilaManager extends TilaManager<VuorovaikutusKierros, 
     return projekti.vuorovaikutusKierros;
   }
 
-  getVaiheAineisto(projekti: DBProjekti): VaiheAineisto<VuorovaikutusKierros, VuorovaikutusKierrosJulkaisu> {
-    return new ProjektiAineistoManager(projekti).getVuorovaikutusKierros();
+  getVaiheAineisto(projekti: DBProjekti): VaiheTiedostoManager<VuorovaikutusKierros, VuorovaikutusKierrosJulkaisu> {
+    return new ProjektiTiedostoManager(projekti).getVuorovaikutusKierros();
   }
 }
